requirements.txt
Flask==3.0.0
PyJWT==2.8.0
cryptography==41.0.0
requests==2.31.0
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
responses==0.24.1
